feat: Enhance Company Certificate Management (CCM) with new notification endpoints and refactor services
